Adding the special command "varregread"
This needs some annoying special command stuff.  Specifically the 2-byte
number which is added to the command spec as "d".  Also, added is a --dev
option on the command line to specify a different device, so you don't
have to recompile if the usb number changes.

Usage:
	vc("varregread", 2, REG_ADDRESS, 2)

You need the 2's to format it right since the command spec assumes (silly us)
that the board has a vaugely consistant format.  Technically the 2's are the
length of the address and the length of the register, but it looks like they're
always 2.

Signed-off-by: Charlie Mooney <charliemooney@chromium.org>