Fixing exit/cleanup code
I noticed that the child process was hanging around after exiting, and
it seemed to be causing problems for me reconnecting to the board.  Now
it kills the child before exiting.

Additionally, SIGINT is now caught by both threads, and after 5 of them
will quit.  This will hopefully keep me from absently hitting Ctrl+C and
killing it messily and loosing my console history, etc, but still lets
us kill it this way if need be.

Signed-off-by: Charlie Mooney <charliemooney@chromium.org>