Fixing pixel parser
With the new info from Bong, I now know that the pixel parsing code was
wrong before.  Each channel is packed like:

           Byte 0            Byte 1
        L L L L L L L L | - - - - - - H H

Where the resulting 10-bit channel value is: HHLLLLLLLL

Each pixel is packed as a series of 3 of those 2-bit channels as:

        R R G G B B

And the pixels are pack in row-wise order.  For instance.  If you
request the block of pixels (0, 0, 3, 2) are they stored like:

p(0, 0) p(1, 0) p(2, 0) p(0, 1) p(1, 1) p(2, 1)

Signed-off-by: Charlie Mooney <charliemooney@chromium.org>