Add function to save .ppm images from frame buf
This adds the saveimage() function to the videochameleon console.  It
takes a chunk of the frame buffer and reads it out, parses it and saves
it to disk in the plain text image file format .ppm.

Usage:  saveimage(filename, x, y, w, h)
   ie:  saveimage("img.ppm", 0, 0, 150, 150)

It will handle multiple pixel requests for you, but only if the width of
the requested section can fit in a single requests.  It won't work if
w > 170!

Signed-off-by: Charlie Mooney <charliemooney@chromium.org>